new announcment added, added tenant delete from room, some toStrings() added

+++ app/app/src/main/java/com/PGmitra/app/Repository/VenderRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface VenderRepo extends JpaRepository<Owner, Integer> {
+public interface VenderRepo extends JpaRepository<Owner, Long> {
     Optional<Owner> findByUsername(String username);
     Optional<Owner> findByEmail(String email);
 }
